double to float

+++ lib/nikfemm/src/geometry/segment.cpp
@@ -9,7 +9,7 @@ namespace nikfemm {
         this->p2 = p2;
     }
 
-    double Segment::length() {
+    float Segment::length() {
         return sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2));
     }
